# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users/{user_id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get User",
        "operationId": "get_user_admin_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete User",
        "operationId": "delete_user_admin_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/avatar/": {
      "post": {
        "tags": [
          "Standalone Avatar"
        ],
        "summary": "Create Standalone Avatar",
        "description": "Create a new avatar independently (before story creation).",
        "operationId": "create_standalone_avatar_avatar__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvatarCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/avatar/{avatar_id}": {
      "get": {
        "tags": [
          "Standalone Avatar"
        ],
        "summary": "Get Standalone Avatar",
        "description": "Retrieve an avatar by ID.",
        "operationId": "get_standalone_avatar_avatar__avatar_id__get",
        "parameters": [
          {
            "name": "avatar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Avatar Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/story-setting/": {
      "post": {
        "tags": [
          "Standalone Story Setting"
        ],
        "summary": "Create Standalone Story Setting",
        "description": "Create a new story setting independently (before story creation).",
        "operationId": "create_standalone_story_setting_story_setting__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorySettingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorySettingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/story/": {
      "get": {
        "tags": [
          "Story"
        ],
        "summary": "List Stories",
        "description": "Get all stories, optionally filtered by user_id.",
        "operationId": "list_stories_story__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoryRead"
                  },
                  "title": "Response List Stories Story  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Story"
        ],
        "summary": "Create Story",
        "description": "Create a new story.",
        "operationId": "create_story_story__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/story/{story_id}/avatar/{avatar_id}": {
      "get": {
        "tags": [
          "Avatar"
        ],
        "summary": "Get Avatar",
        "description": "Retrieve a single avatar by ID for the story.",
        "operationId": "get_avatar_story__story_id__avatar__avatar_id__get",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Id"
            }
          },
          {
            "name": "avatar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Avatar Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Avatar"
        ],
        "summary": "Update Avatar",
        "description": "Update an existing avatar for the story.",
        "operationId": "update_avatar_story__story_id__avatar__avatar_id__put",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Id"
            }
          },
          {
            "name": "avatar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Avatar Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvatarUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AvatarCreate": {
        "properties": {
          "avatar_name": {
            "type": "string",
            "title": "Avatar Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "avatar_name"
        ],
        "title": "AvatarCreate",
        "description": "Creating new avatar character.\nUsed during character creation phase before the story begins.\n\nUser provides avatar name and description; System generates avatar_id and timestamps automatically."
      },
      "AvatarRead": {
        "properties": {
          "avatar_id": {
            "type": "integer",
            "title": "Avatar Id"
          },
          "avatar_name": {
            "type": "string",
            "title": "Avatar Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "avatar_id",
          "avatar_name",
          "created_at"
        ],
        "title": "AvatarRead",
        "description": "Returning avatar information via API.\nUsed when displaying avatar list/avatar details to users."
      },
      "AvatarUpdate": {
        "properties": {
          "avatar_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "AvatarUpdate",
        "description": "Updating avatar information.\nUsed when user wants to rename their avatar or update description.\n\nOptional."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StoryCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "avatar_id": {
            "type": "integer",
            "title": "Avatar Id"
          },
          "story_setting_id": {
            "type": "integer",
            "title": "Story Setting Id"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "avatar_id",
          "story_setting_id"
        ],
        "title": "StoryCreate",
        "description": "Creating a new story."
      },
      "StoryRead": {
        "properties": {
          "story_id": {
            "type": "integer",
            "title": "Story Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "avatar_id": {
            "type": "integer",
            "title": "Avatar Id"
          },
          "story_setting_id": {
            "type": "integer",
            "title": "Story Setting Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "story_id",
          "user_id",
          "avatar_id",
          "story_setting_id",
          "created_at"
        ],
        "title": "StoryRead",
        "description": "Returning story information.\nUsed when displaying story details or resuming a story session."
      },
      "StorySettingCreate": {
        "properties": {
          "setting_prompt": {
            "type": "string",
            "title": "Setting Prompt"
          }
        },
        "type": "object",
        "required": [
          "setting_prompt"
        ],
        "title": "StorySettingCreate",
        "description": "Creating a story setting/prompt."
      },
      "StorySettingRead": {
        "properties": {
          "story_setting_id": {
            "type": "integer",
            "title": "Story Setting Id"
          },
          "setting_prompt": {
            "type": "string",
            "title": "Setting Prompt"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "story_setting_id",
          "setting_prompt",
          "created_at"
        ],
        "title": "StorySettingRead",
        "description": "Returning story setting information.\nUsed when displaying story configuration/setting details."
      },
...
      "UserCreate": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "UserCreate",
        "description": "Creating a new user account.\nUsed when registering a new user in the system."
      },
      "UserLogin": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserLogin",
        "description": "Login credentials."
      },
      "UserRead": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "access_level": {
            "type": "string",
            "title": "Access Level"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "username",
          "email",
          "access_level",
          "created_at"
        ],
        "title": "UserRead",
        "description": "Returning user for information via API."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}